SmartLabManager/SmartLabManager/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "streamlit-option-menu>=0.4.0",
    "streamlit>=1.43.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "py4j>=0.10.9.9",
    "pillow-avif-plugin>=1.5.1",
]
